requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
bcrypt==4.0.1
cachetools==5.3.1
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
fastapi-cache2==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
pendulum==2.1.2
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-decouple==3.8
pytzdata==2020.1
redis==4.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing==3.7.4.3
typing_extensions==4.8.0
uvicorn==0.23.2
